+++ Sources/SentryObjCCompat/SentryObjCBreadcrumb.swift
@@ -53,6 +53,7 @@ import Foundation
 
     @objc public var data: [String: Any]? {
         get { wrapped.data }
+        @available(*, deprecated, message: "Use setData(value:key:) instead.")
         // swiftlint:disable no_breadcrumb_data_setter
         set { wrapped.data = newValue } // pass-through mirroring the data setter; removed in V10
         // swiftlint:enable no_breadcrumb_data_setter